@@ -7,7 +7,7 @@ namespace PanoramicData.SheetMagic.Test
 	public abstract class Test
 	{
 		protected static FileInfo GetXlsxTempFileInfo()
-	 		=> new FileInfo(Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid()}.xlsx"));
+	 		=> new(Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid()}.xlsx"));
 
 		protected static FileInfo GetSheetFileInfo(string worksheetName)
 		{
